Publish a developer package by GitHub Actions (#25)
* feat(gha): build package

Introduces the following workflows:
- `publish-dev-package`: publishes a developer package when commits are
  pushed to the `dev` branch. Invokes the `publish-package` workflow.
- `publish-package`: implements the process to publish a package.
  Designed for the GitHub npm registry for now.

* feat(gha): append short commit hash to version

The `publish-package` workflow appends the short commit hash to the
version for a developer package.

* feat(gha): publish developer package

The `publish-package` workflow publishes the built package to a
specified npm registry. The caller has to specify the following
parameters:
- `npm-registry-url`: URL of the npm registry to publish to.
- `npm-token`: token that is allowed to publish to the npm registry

The `publish-dev-package` workflow configures the above parameters for
the GitHub npm registry.

* fix(gha): unclean working tree error on publish

Works around the problem that the `pnpm publish` command failed due to
an unclean git working copy. As the short commit hash is appended to the
version of a developer package, the working copy is always unclean. Adds
the `--no-git-checks` option to the `pnpm publish` command as a
workaround.

* chore: include README.ja.md in package

* chore(gha): specify the scope of the package

* chore: add repository to package.json

* chore(gha): change trigger branch: dev → main

The `publish-dev-package` workflow changes the trigger branch from `dev`
to `main`. I decided not to duplicate the `main` branch as `dev`.

When it would come to publishing to the official npm registry, we should
consider other triggers: e.g. creation of a release.

* docs: update README

Explains how to install a developer package from the GitHub npm
registry.